refactor: cascade cleanup wave 4 — GREEN BUILD (43 remaining files)
All 43 remaining cascade-affected files brought green via 3 parallel
cleanup passes:

Certificates/ (15 files): AuthCommitteeHot, PoolRegistration, Register,
RegisterDRep, ResignCommitteeCold, StakeDelegation, StakeDeregistration,
StakeRegisterDelegate, StakeRegistration, StakeVoteDelegate,
StakeVoteRegisterDelegate, Unregister, UnregisterDRep, UpdateDRep,
VoteDelegate. Pattern: .uint(UInt(CODE.rawValue)) → .uint(UInt64(...));
.int(Int(coin)) → .int(Int64(coin)).

Governance/ + NativeScript/ + Plutus/ (12 files): Committee, Governance,
HardForkInitiationAction, InfoAction, NewConstitution, NoConfidence,
ProposalProcedure, ScriptAll, ScriptAny, ScriptPubkey, BigInteger,
RawPlutusData. ProposalProcedure.deposit (Coin) no longer needs the
UInt(...) wrap; BigInteger drops redundant Int() narrowing on its
already-Int64 associated value.

Block/ + Credentials/ + Crypto/ + Legacy/ + Serialization/ + StakePool/
+ Transaction/ + Types/ (16 files): EraBlock, EraBlockHeader, Credential,
Bip32, GenesisKeyDelegation, Legacy/Update, JSONSerializable,
MultiHostName, Script, TransactionInput, UTxO, Withdrawals, CBORTag,
Fraction, PositiveCoin, ProtocolVersion. Bip32 path-parsing now uses
UInt32 directly (matches the derive(index:) signature widened earlier).
Fraction.integerValue() generic resolution needed typed locals.

**`swift build` is now green from a clean working tree.** Total cascade
resolved this session: 15,046 → 0 errors across 85 files. Test run,
Linux verification, and watchOS xcodebuild are next steps.

+++ Sources/SwiftCardanoCore/Block/EraBlock.swift
@@ -492,7 +492,7 @@ extension EraBlock: Serializable {
             innerBytes = try b.toCBORData()
         }
         return .list([
-            .uint(UInt(eraId)),
+            .uint(UInt64(eraId)),
             .cborTag(CBORTag(tag: 24, value: .bytes(innerBytes))),
         ])
     }
@@ -531,7 +531,7 @@ extension EraBlock: Serializable {
 
     public func toDict() throws -> Primitive {
         var dict = OrderedDictionary<Primitive, Primitive>()
-        dict[.string("era")] = .uint(UInt(eraId))
+        dict[.string("era")] = .uint(UInt64(eraId))
         let blockP: Primitive
         switch self {
         case .byron(let b): blockP = try b.toDict()
