+++ Sources/SwiftCardanoCore/Block/EraBlockHeader.swift
@@ -452,7 +452,7 @@ extension EraBlockHeader: Serializable {
             innerBytes = try h.toCBORData()
         }
         return .list([
-            .uint(UInt(eraId)),
+            .uint(UInt64(eraId)),
             .cborTag(CBORTag(tag: 24, value: .bytes(innerBytes))),
         ])
     }
@@ -492,7 +492,7 @@ extension EraBlockHeader: Serializable {
 
     public func toDict() throws -> Primitive {
         var dict = OrderedDictionary<Primitive, Primitive>()
-        dict[.string("era")] = .uint(UInt(eraId))
+        dict[.string("era")] = .uint(UInt64(eraId))
         let headerP: Primitive
         switch self {
         case .byron(let h): headerP = try h.toDict()
