+++ Sources/SwiftCardanoCore/Certificates/AuthCommitteeHot.swift
@@ -81,7 +81,7 @@ public struct AuthCommitteeHot: CertificateSerializable {
 
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try committeeColdCredential.toPrimitive(),
             try committeeHotCredential.toPrimitive()
         ])
