+++ Sources/SwiftCardanoCore/Certificates/RegisterDRep.swift
@@ -94,9 +94,9 @@ public struct RegisterDRep: CertificateSerializable {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try drepCredential.toPrimitive(),
-            .int(Int(coin)),
+            .int(Int64(coin)),
             try anchor?.toPrimitive() ?? .null
         ])
     }
@@ -134,7 +134,7 @@ public struct RegisterDRep: CertificateSerializable {
     public func toDict() throws -> Primitive {
         var dict = OrderedDictionary<Primitive, Primitive>()
         dict[.string(CodingKeys.drepCredential.rawValue)] = .string(try drepCredential.id())
-        dict[.string(CodingKeys.coin.rawValue)] = .int(Int(coin))
+        dict[.string(CodingKeys.coin.rawValue)] = .int(Int64(coin))
         if let anchor = anchor {
             dict[.string(CodingKeys.anchor.rawValue)] = try anchor.toDict()
         } else {
