+++ Sources/SwiftCardanoCore/Certificates/StakeDeregistration.swift
@@ -70,7 +70,7 @@ public struct StakeDeregistration: CertificateSerializable {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try stakeCredential.toPrimitive()
         ])
     }
