+++ Sources/SwiftCardanoCore/Certificates/StakeRegistration.swift
@@ -72,7 +72,7 @@ public struct StakeRegistration: CertificateSerializable {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try stakeCredential.toPrimitive()
         ])
     }
