+++ Sources/SwiftCardanoCore/Certificates/StakeVoteDelegate.swift
@@ -87,7 +87,7 @@ public struct StakeVoteDelegate: CertificateSerializable {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try stakeCredential.toPrimitive(),
             poolKeyHash.toPrimitive(),
             try drep.toPrimitive()
