+++ Sources/SwiftCardanoCore/Certificates/StakeVoteRegisterDelegate.swift
@@ -102,11 +102,11 @@ public struct StakeVoteRegisterDelegate: CertificateSerializable {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .uint(UInt(Self.CODE.rawValue)),
+            .uint(UInt64(Self.CODE.rawValue)),
             try stakeCredential.toPrimitive(),
             poolKeyHash.toPrimitive(),
             try drep.toPrimitive(),
-            .int(Int(coin))
+            .int(Int64(coin))
         ])
     }
     
@@ -154,7 +154,7 @@ public struct StakeVoteRegisterDelegate: CertificateSerializable {
         dict[.string(CodingKeys.stakeCredential.rawValue)] = .string(stakeCredential.credential.payload.toHex)
         dict[.string(CodingKeys.poolKeyHash.rawValue)] = .string(try poolOperator.id())
         dict[.string(CodingKeys.drep.rawValue)] = .string(try drep.id())
-        dict[.string(CodingKeys.coin.rawValue)] = .int(Int(coin))
+        dict[.string(CodingKeys.coin.rawValue)] = .int(Int64(coin))
         
         return .orderedDict(dict)
     }
