+++ Sources/SwiftCardanoCore/Governance/GovernanceActions/NewConstitution.swift
@@ -37,7 +37,7 @@ public struct NewConstitution: GovernanceAction {
         }
         let code: Int
         switch elements[0] {
-        case .int(let v): code = v
+        case .int(let v): code = Int(v)
         case .uint(let v): code = Int(v)
         default: throw CardanoCoreError.deserializeError("Invalid NewConstitution primitive")
         }
@@ -51,7 +51,7 @@ public struct NewConstitution: GovernanceAction {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .int(Self.code.rawValue),
+            .int(Int64(Self.code.rawValue)),
             try id.toPrimitive(),
             try constitution.toPrimitive()
         ])
