+++ Sources/SwiftCardanoCore/Governance/GovernanceActions/NoConfidence.swift
@@ -33,7 +33,7 @@ public struct NoConfidence: GovernanceAction {
         }
         let code: Int
         switch elements[0] {
-        case .int(let v): code = v
+        case .int(let v): code = Int(v)
         case .uint(let v): code = Int(v)
         default: throw CardanoCoreError.deserializeError("Invalid NoConfidence primitive")
         }
@@ -45,7 +45,7 @@ public struct NoConfidence: GovernanceAction {
     
     public func toPrimitive() throws -> Primitive {
         return .list([
-            .int(Self.code.rawValue),
+            .int(Int64(Self.code.rawValue)),
             try id.toPrimitive()
         ])
     }
