+++ Sources/SwiftCardanoCore/Transaction/UTxO.swift
@@ -47,7 +47,7 @@ public struct UTxO: Serializable {
                 outputPrimitives: (String, Int, DatumOption?, ScriptType?, Bool?)) throws {
         input = try TransactionInput(from: inputPrimitives.0, index: inputPrimitives.1)
         output = try TransactionOutput(from: outputPrimitives.0,
-                                   amount: outputPrimitives.1,
+                                   amount: Int64(outputPrimitives.1),
                                    datumOption: outputPrimitives.2,
                                    script: outputPrimitives.3,
                                    postAlonzo: outputPrimitives.4 ?? true)
